requirements.txt
python-dotenv
pandas
numpy
pymango
certifi
pymongo[srv]
dill
scikit-lean
pyaml
mlflow
#-e .